@@ -4,68 +4,78 @@
   "description": "",
   "main": "server.js",
   "dependencies": {
-    "babel": "^6.3.26",
-    "babel-polyfill": "^6.3.14",
-    "deap": "^1.0.0",
-    "good": "~6.4.0",
-    "good-console": "^5.2.0",
-    "gulp-develop-server": "^0.5.0",
-    "gulp-uglify": "^1.5.1",
-    "hapi": "^12.1.0",
-    "highlight.js": "^9.1.0",
-    "history": "^2.0.0-rc2",
-    "immutable": "^3.7.6",
-    "inert": "^3.2.0",
+    "@hapi/good": "^8.2.4",
+    "@hapi/good-console": "^8.1.2",
+    "@hapi/good-squeeze": "^5.2.1",
+    "@hapi/inert": "^5.2.2",
+    "@hapi/vision": "^5.5.4",
+    "good": "~8.1.2",
+    "good-console": "^8.0.0",
+    "gulp-develop-server": "^0.5.2",
+    "gulp-uglify": "^3.0.2",
+    "hapi": "^18.1.0",
+    "highlight.js": "^9.15.10",
+    "history": "^4.10.1",
+    "immutable": "^3.8.2",
+    "inert": "^5.1.3",
     "jade": "^1.11.0",
-    "jsonparse": "^1.2.0",
-    "marked": "0.6.2",
-    "md5": "^2.0.0",
-    "nodemailer": "2.0.0",
-    "phantom": "^0.8.4",
-    "react": "^0.14.6",
-    "react-dom": "^0.14.6",
-    "react-router": "^2.0.0-rc5",
-    "recast": "^0.11.0",
-    "rss": "^1.2.1",
-    "sitemap": "^1.5.0",
-    "striptags": "~2.1.1",
+    "jsonparse": "^1.3.1",
+    "marked": "0.7.0",
+    "md5": "^2.2.1",
+    "nodemailer": "6.3.1",
+    "phantom": "^6.3.0",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "react-router-dom": "^5.1.2",
+    "recast": "^0.18.5",
+    "rss": "^1.2.2",
+    "sitemap": "^5.0.1",
+    "striptags": "~3.1.1",
     "through": "^2.3.8",
-    "through2": "^2.0.0",
-    "uglify-js": "^2.6.1",
-    "underscore": "^1.8.3",
-    "vinyl-buffer": "^1.0.0",
+    "through2": "^3.0.1",
+    "uglify-js": "^3.6.4",
+    "underscore": "^1.9.1",
+    "vinyl-buffer": "^1.0.1",
     "vinyl-sourcemaps-apply": "^0.2.1",
-    "vision": "^4.0.1",
-    "xhr": "^2.2.0",
-    "yar": "^6.0.0"
+    "vision": "^5.4.4",
+    "xhr": "^2.5.0"
   },
   "devDependencies": {
-    "babel-core": "^6.4.5",
-    "babel-jest": "^6.0.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babelify": "^7.2.0",
-    "browserify": "^13.0.0",
-    "del": "^2.2.0",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^6.1.1",
-    "gulp-changed": "^1.3.0",
-    "gulp-less": "^3.0.5",
-    "gulp-nodemon": "^2.0.6",
-    "gulp-shell": "^0.5.1",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-standard": "^5.2.2",
-    "gulp-util": "^3.0.7",
-    "jest-cli": "^0.8.2",
-    "nodemon": "^1.8.1",
-    "pre-commit": "^1.1.2",
-    "run-sequence": "^1.1.5",
-    "standard": "^5.4.1",
-    "vinyl-source-stream": "^1.1.0"
+    "@babel/core": "^7.6.4",
+    "@babel/polyfill": "^7.6.0",
+    "@babel/preset-env": "^7.6.3",
+    "@babel/preset-react": "^7.6.3",
+    "acorn": "^7.1.0",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babelify": "^10.0.0",
+    "browserify": "^16.5.0",
+    "del": "^5.1.0",
+    "eslint": "^6.6.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^10.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.16.0",
+    "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
+    "gulp-changed": "^4.0.2",
+    "gulp-less": "^4.0.1",
+    "gulp-nodemon": "^2.4.2",
+    "gulp-shell": "^0.7.1",
+    "gulp-sourcemaps": "^2.6.5",
+    "gulp-standard": "^14.0.0",
+    "gulp-util": "^3.0.8",
+    "jest-cli": "^24.9.0",
+    "nodemon": "^1.19.4",
+    "pre-commit": "^1.2.2",
+    "run-sequence": "^2.2.1",
+    "standard": "^14.3.1",
+    "vinyl-source-stream": "^2.0.0"
   },
   "scripts": {
     "start": "gulp",
-    "lint": "gulp lint",
+    "lint": "standard",
     "test": "jest dist"
   },
   "repository": {
@@ -91,7 +101,8 @@
   },
   "homepage": "https://github.com/bbondy/brianbondy.node#readme",
   "standard": {
-    "ignore": []
+    "ignore": [],
+    "parser": "babel-eslint"
   },
   "pre-commit": [
     "lint"
