@@ -1,19 +1,19 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
 import Tag from './tag.js'
-import {fetchBlogPost} from './client.js'
-import {formatDate} from './formatDate.js'
+import { fetchBlogPost } from './client.js'
+import { formatDate } from './formatDate.js'
 import externalLinkSetup from './externalLinkSetup.js'
 
 export default class BlogPost extends React.Component {
   constructor () {
     super()
+    this.blogDiv = React.createRef()
     this.state = {}
   }
 
-  componentWillMount () {
-    if (this.props.params && this.props.params.id) {
-      fetchBlogPost(this.props.params.id).then((blogPost) => {
+  componentDidMount () {
+    if (this.props.match && this.props.match.params && this.props.match.params.id) {
+      fetchBlogPost(this.props.match.params.id).then((blogPost) => {
         document.title = blogPost.title + ' - Brian R. Bondy'
         this.setState({
           id: blogPost.id,
@@ -26,14 +26,14 @@ export default class BlogPost extends React.Component {
         return blogPost
       })
     } else {
-      this.state = {
+      this.setState({
         id: this.props.id,
         title: this.props.title,
         body: this.props.body,
         created: new Date(this.props.created),
         tags: this.props.tags,
         url: this.props.url
-      }
+      })
     }
   }
 
@@ -42,20 +42,19 @@ export default class BlogPost extends React.Component {
   }
 
   componentDidUpdate () {
-    externalLinkSetup(ReactDOM.findDOMNode(this.refs.blogDiv))
+    externalLinkSetup(this.blogDiv.current)
   }
 
   render () {
     if (!this.state.title) {
       return null
     }
-    return <article className='blogPost'>
-      <h1><a href={this.blogPostURL}>{this.state.title}</a></h1>
-      <div className='datePosted'>Posted on {formatDate(this.state.created)}</div>
-      <div ref='blogDiv' dangerouslySetInnerHTML={{__html: this.state.body}}/>
-      <div className='tags'>{this.state.tags.map(tag => <Tag key={tag} name={tag}/>)}</div>
-    </article>
+    return (
+      <article className='blogPost'>
+        <h1><a href={this.blogPostURL}>{this.state.title}</a></h1>
+        <div className='datePosted'>Posted on {formatDate(this.state.created)}</div>
+        <div ref={this.blogDiv} dangerouslySetInnerHTML={{ __html: this.state.body }} />
+        <div className='tags'>{this.state.tags.map(tag => <Tag key={tag} name={tag} />)}</div>
+      </article>)
   }
 }
-
-BlogPost.propTypes = { id: React.PropTypes.any, url: React.PropTypes.string, tags: React.PropTypes.arrayOf(React.PropTypes.string), created: React.PropTypes.any, body: React.PropTypes.string, title: React.PropTypes.string, params: React.PropTypes.object }
