@@ -1,13 +1,13 @@
-var _ = require('underscore')
+import { blogPosts } from './cache.js'
 
-import {blogPosts} from './cache.js'
+var _ = require('underscore')
 
 export function setupRedirects (server) {
   server.route({
     method: 'GET',
     path: '/blog/modified/{year}',
     handler: function (request, reply) {
-      reply.redirect(`/blog/posted/${request.params.year}`).permanent()
+      return reply.redirect(`/blog/posted/${request.params.year}`).permanent()
     }
   })
 
@@ -15,9 +15,9 @@ export function setupRedirects (server) {
     method: 'GET',
     path: '/blog/id/{id}/{slug}',
     handler: function (request, reply) {
-      let blogPost = _(blogPosts).find(post =>
+      const blogPost = _(blogPosts).find(post =>
         post.id === Number(request.params.id))
-      reply.redirect(blogPost.url).permanent()
+      return reply.redirect(blogPost.url).permanent()
     }
   })
 
@@ -25,7 +25,7 @@ export function setupRedirects (server) {
     method: 'GET',
     path: '/blog/page/{page}',
     handler: function (request, reply) {
-      reply.redirect(`/page/${request.params.page}`).permanent()
+      return reply.redirect(`/page/${request.params.page}`).permanent()
     }
   })
 
@@ -33,9 +33,9 @@ export function setupRedirects (server) {
     method: 'GET',
     path: '/blog/id/{id}',
     handler: function (request, reply) {
-      let blogPost = _(blogPosts).find(post =>
+      const blogPost = _(blogPosts).find(post =>
         post.id === Number(request.params.id))
-      reply.redirect(blogPost.url).permanent()
+      return reply.redirect(blogPost.url).permanent()
     }
   })
 
@@ -43,9 +43,9 @@ export function setupRedirects (server) {
     method: 'GET',
     path: '/blog/{id}',
     handler: function (request, reply) {
-      let blogPost = _(blogPosts).find(post =>
+      const blogPost = _(blogPosts).find(post =>
         post.id === Number(request.params.id))
-      reply.redirect(blogPost.url).permanent()
+      return reply.redirect(blogPost.url).permanent()
     }
   })
 
@@ -53,14 +53,14 @@ export function setupRedirects (server) {
     method: 'GET',
     path: '/mozilla/cheatsheet',
     handler: function (request, reply) {
-      reply.redirect('http://www.codefirefox.com/cheatsheet').permanent()
+      return reply.redirect('http://www.codefirefox.com/cheatsheet').permanent()
     }
   })
 
   // Note that I'm intentionally not using the stripTrailingSlash: true option because it
   // simply serves a second endpoint which is bad for PR juice.  301 is better.
   const maxDepth = 6
-  let handler = (request, reply) => reply.redirect(`/${request.params.p}`).permanent()
+  const handler = (request, reply) => reply.redirect(`/${request.params.p}`).permanent()
   for (var i = 1; i <= maxDepth; i++) {
     server.route({
       method: 'GET',
