@@ -1,4 +1,4 @@
-var _gaq = _gaq || []
+var _gaq = window._gaq || []
 _gaq.push(['_setAccount', 'UA-575471-1'])
 _gaq.push(['_trackPageview'])
 
@@ -6,7 +6,7 @@ _gaq.push(['_trackPageview'])
   var ga = document.createElement('script')
   ga.type = 'text/javascript'
   ga.async = true
-  ga.src = ('https:' == document.location.protocol
+  ga.src = (document.location.protocol === 'https:'
     ? 'https://ssl' : 'http://www') + '.google-analytics.com/ga.js'
   var s = document.getElementsByTagName('script')[0]
   s.parentNode.insertBefore(ga, s)
