@@ -1,21 +1,21 @@
 export function titleByPath (pathname) {
-  let parts = pathname.substring(1).split('/')
+  const parts = pathname.substring(1).split('/')
   if (pathname.startsWith('/blog/tagged/')) {
-    let tag = parts[2]
+    const tag = parts[2]
     let page = ''
     if (parts[3] === 'page') {
       page = ` page ${parts[4]}`
     }
     return `Posts tagged ${tag}${page}`
   } else if (pathname.startsWith('/blog/posted/')) {
-    let year = parts[2]
+    const year = parts[2]
     let page = ''
     if (parts[3] === 'page') {
       page = ` page ${parts[4]}`
     }
     return `Posted in ${year}${page}`
   } else if (pathname.startsWith('/page/')) {
-    let page = parts[1]
+    const page = parts[1]
     return `Blog posts page ${page}`
   }
 
@@ -57,8 +57,8 @@ export function titleByPath (pathname) {
 }
 
 export function pageTitleByPath (pathname) {
-  let suffix = 'Brian R. Bondy'
-  let title = titleByPath(pathname)
+  const suffix = 'Brian R. Bondy'
+  const title = titleByPath(pathname)
   if (!title) {
     return suffix
   }
